# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.AdminApplicationApprovalResponse:
    properties:
      applicationId:
        example: app-555666
        type: string
      roleAssigned:
        example: artist
        type: string
      status:
        example: approved
        type: string
      updatedAt:
        type: string
      userId:
        example: user-keycloak-uuid
        type: string
    type: object
  handlers.AdminPingResponse:
    properties:
      message:
        example: admin pong
        type: string
    type: object
  handlers.AdminStatsResponse:
    properties:
      activeOrders:
        example: 12
        type: integer
      totalRevenue:
        example: 1450.5
        type: number
      totalUsers:
        example: 154
        type: integer
    type: object
  handlers.ArtObjectResponse:
    properties:
      artistId:
        example: artist-4321
        type: string
      basePrice:
        example: 5000
        type: number
      country:
        example: CZ
        type: string
      createdAt:
        type: string
      currency:
        example: EUR
        type: string
      currentHolderId:
        example: partner-9876
        type: string
      currentPrice:
        example: 5075.5
        type: number
      dailyGrowthRate:
        example: 0.0001
        type: number
      id:
        example: 64a7b3e1f1d2c3a4b5
        type: string
      title:
        example: Sunset over Prague
        type: string
    type: object
  handlers.BuyShareInput:
    properties:
      objectId:
        example: 64a7b3e1f1d2c3a4b5
        type: string
      sharePct:
        example: 5
        maximum: 10
        minimum: 1
        type: number
    required:
    - objectId
    - sharePct
    type: object
  handlers.CreatorApplicationInput:
    properties:
      bio:
        example: Contemporary oil painter based in Prague, focusing on urban landscapes.
        maxLength: 500
        type: string
      portfolioUrl:
        example: https://behance.net/my-art-portfolio
        type: string
      socialLinks:
        example:
        - '[''https://instagram.com/artist_handle'']'
        items:
          type: string
        type: array
    required:
    - bio
    - portfolioUrl
    - socialLinks
    type: object
  handlers.CreatorApplicationResponse:
    properties:
      applicationId:
        example: app-111222
        type: string
      bio:
        type: string
      createdAt:
        type: string
      portfolioUrl:
        type: string
      status:
        description: pending, approved, rejected
        example: pending
        type: string
      userId:
        example: user-keycloak-uuid
        type: string
    type: object
  handlers.TransactionResponse:
    properties:
      amountPaid:
        example: 253.75
        type: number
      buyerId:
        example: user-buyer-uuid
        type: string
      createdAt:
        type: string
      currency:
        example: EUR
        type: string
      objectId:
        example: 64a7b3e1f1d2c3a4b5
        type: string
      sellerId:
        example: artist-seller-uuid
        type: string
      sharePct:
        example: 5
        type: number
      transactionId:
        example: tx-abc-999888
        type: string
    type: object
  handlers.TransferApprovalResponse:
    properties:
      approvedAt:
        type: string
      status:
        example: approved
        type: string
      transferId:
        example: tr-888999
        type: string
    type: object
  handlers.TransferRequestInput:
    properties:
      durationDays:
        example: 7
        minimum: 1
        type: integer
      eventDetails:
        example: Corporate pop-up exhibition or apartment art party
        type: string
      objectId:
        example: 64a7b3e1f1d2c3a4b5
        type: string
    required:
    - durationDays
    - eventDetails
    - objectId
    type: object
  handlers.TransferRequestResponse:
    properties:
      createdAt:
        type: string
      eventDetails:
        example: Corporate pop-up exhibition or apartment art party
        type: string
      expiresAt:
        type: string
      fromHolderId:
        example: partner-current-owner-uuid
        type: string
      objectId:
        example: 64a7b3e1f1d2c3a4b5
        type: string
      requesterId:
        example: partner-host-uuid
        type: string
      status:
        description: pending, approved, rejected, completed
        example: pending
        type: string
      transferId:
        example: tr-999111
        type: string
    type: object
  handlers.UserProfileResponse:
    properties:
      email:
        example: user@artofino.com
        type: string
      firstName:
        example: John
        type: string
      id:
        example: 4321-abcd-1234
        type: string
      lastName:
        example: Doe
        type: string
      username:
        example: johndoe
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: Backend service for ArtoFino platform.
  title: ArtoFino Backend API
  version: "1.0"
paths:
  /admin/applications/{id}/approve:
    post:
      description: Allows an administrator to review and approve a pending creator
        application, upgrading the user's system role.
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AdminApplicationApprovalResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve an artist application
      tags:
      - admin
  /admin/ping:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AdminPingResponse'
      security:
      - BearerAuth: []
      summary: Admin ping
      tags:
      - admin
  /admin/stats:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AdminStatsResponse'
      security:
      - BearerAuth: []
      summary: Admin stats
      tags:
      - admin
  /authors/apply:
    post:
      consumes:
      - application/json
      description: Allows a registered user to submit portfolio and social links to
        get vetted and upgraded to the 'artist' role.
      parameters:
      - description: Artist application data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatorApplicationInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.CreatorApplicationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Apply for Artist (Creator) status
      tags:
      - authors
  /hc:
    get:
      description: Returns OK if service is alive
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Health check
      tags:
      - system
  /lc:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: ALIVE
          schema:
            type: string
      summary: Liveness check
      tags:
      - system
  /objects/{id}:
    get:
      description: Returns art object data with current price calculated based on
        daily linear growth, custom currency and dynamic rate
      parameters:
      - description: Object ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ArtObjectResponse'
      summary: Get art object with dynamic price
      tags:
      - objects
  /rc:
    get:
      produces:
      - text/plain
      responses:
        "200":
          description: READY
          schema:
            type: string
        "503":
          description: NOT READY
          schema:
            type: string
      summary: Readiness check
      tags:
      - system
  /transactions/buy:
    post:
      consumes:
      - application/json
      description: Purchase between 1% and 10% of an art object. Price is locked at
        the current calculated rate using the linear growth formula.
      parameters:
      - description: Purchase data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.BuyShareInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Buy a share in an art object
      tags:
      - transactions
  /transfers/{id}/approve:
    post:
      description: Allows the current holder or owner of the art object to approve
        a pending temporary transfer request.
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TransferApprovalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a physical transfer request
      tags:
      - transfers
  /transfers/request:
    post:
      consumes:
      - application/json
      description: Allows a Partner to apply for hosting an art object at their event/location.
      parameters:
      - description: Transfer request details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.TransferRequestInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TransferRequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Request art object for temporary use
      tags:
      - transfers
  /users/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"